+++ plugins/org.jkiss.dbeaver.ext.db2.ui/META-INF/MANIFEST.MF
@@ -1,0 +1,23 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: %Bundle-Name
+Bundle-SymbolicName: org.jkiss.dbeaver.ext.db2.ui;singleton:=true
+Bundle-Version: 1.0.1.qualifier
+Bundle-Release-Date: 20200504
+Bundle-Activator: org.jkiss.dbeaver.ext.db2.ui.Activator
+Require-Bundle: org.eclipse.ui,
+ org.jkiss.dbeaver.ui.editors.connection,
+ org.jkiss.dbeaver.model,
+ org.jkiss.dbeaver.ui,
+ org.jkiss.dbeaver.ext.db2,
+ org.jkiss.dbeaver.ui.editors.data,
+ org.jkiss.dbeaver.ui.editors.acl,
+ org.jkiss.dbeaver.ui.editors.sql,
+ org.jkiss.dbeaver.tasks.ui
+Bundle-ActivationPolicy: lazy
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Bundle-Vendor: %Bundle-Vendor
+Bundle-Localization: OSGI-INF/l10n/bundle
+Automatic-Module-Name: org.jkiss.dbeaver.ext.db2.ui
+Import-Package: org.jkiss.dbeaver.ui.editors.object.struct
+Export-Package: org.jkiss.dbeaver.ext.db2.ui.editors
