+++ plugins/org.jkiss.dbeaver.ext.db2.ui/src/org/jkiss/dbeaver/ext/db2/ui/internal/DB2Resources.properties
@@ -1,0 +1,146 @@
+db2_connection_page_tab_database=Database
+db2_connection_page_tab_security=Security
+db2_connection_trace_page_checkbox_connect=Connects
+db2_connection_trace_page_checkbox_connection_calls=Connection calls
+db2_connection_trace_page_checkbox_diagnostics=Diagnostics
+db2_connection_trace_page_checkbox_drda_flows=DRDA flows
+db2_connection_trace_page_checkbox_driver_configuration=Driver configuration
+db2_connection_trace_page_checkbox_enable_trace=Enable trace
+db2_connection_trace_page_checkbox_parameter_metadata=Parameter metadata
+db2_connection_trace_page_checkbox_result_set_calls=Result set calls
+db2_connection_trace_page_checkbox_result_set_metadata=Result set metadata
+db2_connection_trace_page_checkbox_sql_j=SQL J
+db2_connection_trace_page_checkbox_statement_calls=Statement calls
+db2_connection_trace_page_checkbox_xa_calls=XA calls
+db2_connection_trace_page_header_levels=Levels
+db2_connection_trace_page_label_file_name=File name
+db2_connection_trace_page_label_folder=Folder
+db2_connection_trace_page_string_trace=trace
+db2_connection_trace_page_tab_description_trace_settings=Trace settings
+db2_connection_trace_page_tab_trace_settings=Trace settings
+
+db2_connection_trace_page_checkbox_append=Append
+dialog_connection_browse_button=Browse ... 
+dialog_connection_database_schema_label=Database/Schema:
+dialog_connection_db_file_chooser_text=Choose database file
+dialog_connection_db_folder_chooser_message=Choose folder with database files
+dialog_connection_db_folder_chooser_text=Choose database folder
+dialog_connection_edit_driver_button=Edit Driver Settings
+dialog_connection_general_tab=General
+dialog_connection_general_tab_tooltip=General connection properties
+dialog_connection_host_label=Host:
+dialog_connection_jdbc_url_=JDBC URL:
+dialog_connection_password_label=Password:
+dialog_connection_path_label=Path:
+dialog_connection_port_label=Port:
+dialog_connection_server_label=Server:
+dialog_connection_test_connection_button=Test Connection ... 
+dialog_connection_user_name_label=User name:
+
+dialog_connection_advanced_tab=Advanced
+dialog_connection_advanced_tab_tooltip=Advanced/custom driver properties
+dialog_connection_basic_tab=Basic
+dialog_connection_connection_type_group=Connection Type
+dialog_connection_custom_tab=Custom
+dialog_connection_host=Host
+dialog_connection_ora_home=Client
+dialog_connection_os_authentication=OS Authentication
+dialog_connection_password=Password
+dialog_connection_port=Port
+dialog_connection_role=Role
+dialog_connection_security_group=Security
+dialog_connection_select_ora_home_msg=Select DB2 home
+dialog_connection_sid_service=SID/Service
+dialog_connection_sid=SID
+dialog_connection_service=Service Name
+dialog_connection_database=Database
+dialog_connection_test_connection=Test Connection ... 
+dialog_connection_tns_tab=TNS
+dialog_connection_user_name=User name
+dialog_connection_ver=v.
+
+edit_db2_constraint_manager_dialog_title=Create constraint
+edit_db2_data_type_manager_dialog_title=Package
+edit_db2_foreign_key_manager_dialog_title=Create foreign key
+edit_db2_index_manager_dialog_title=Create index
+edit_db2_package_manager_dialog_title=Package
+edit_db2_schema_manager_dialog_title=Schema name
+edit_db2_sequence_manager_dialog_title=Sequence
+edit_db2_trigger_manager_dialog_title=Trigger
+
+editors_db2_application_editor_title_force_application=Force Application
+editors_db2_application_editor_action_force=Force
+editors_db2_application_editor_confirm_action=Are you sure you want to "force" application "{1}"?
+
+dialog_explain_choose_tablespace=Choose a Tablespace
+dialog_explain_choose_tablespace_tablespace=Tablespace
+dialog_explain_no_tables=Explain tables missing
+dialog_explain_no_tablespace_found=No usable Tablespace has been found to create Explain tables
+dialog_explain_no_tablespace_found_title=No Tablespace
+dialog_explain_no_tables_found_ex=EXPLAIN tables not found. Query cannot be explained.
+dialog_explain_ask_to_create=Tables for EXPLAIN not found within current authorization ID (%s) nor in SYSTOOLS.\n\nDo you want DBeaver to create new EXPLAIN tables?
+
+dialog_table_tools_success_title=Success
+dialog_table_tools_options=Options
+dialog_table_tools_result=Result
+dialog_table_tools_progress=Progress
+
+dialog_table_tools_runstats_title=Runstats on Table
+dialog_table_tools_runstats_cols_title=Collect Colums Statistics
+dialog_table_tools_runstats_cols_all=All Columns
+dialog_table_tools_runstats_cols_all_and_distribution=All Columns with Distribution 
+dialog_table_tools_runstats_cols_no=No Statistics on Columns
+dialog_table_tools_runstats_indexes_title=Collect Indexes Statistics
+dialog_table_tools_runstats_indexes_detailed=Extended Detailed Statistics with Sampling on all Indexes
+dialog_table_tools_runstats_indexes_all=All Indexes
+dialog_table_tools_runstats_indexes_no=No Statistics on Indexes
+dialog_table_tools_runstats_stats_title=Table Sampling (% of Pages)
+
+dialog_table_tools_reorg_title=Reorg Table
+dialog_table_tools_reorg_inplace=Reorg in place?
+dialog_table_tools_reorg_useindex=Reorg Using Existing Index?
+dialog_table_tools_reorg_useindexscan=Use Index Scan?
+dialog_table_tools_reorg_truncate=Truncate Table After Reorg?
+dialog_table_tools_reorg_usetempts=Use Temporary Tablespace?
+dialog_table_tools_reorg_reorglobs=Reorg LOBs and Long Columns?
+dialog_table_tools_reorg_reorglobsts=Use Temp. Tablespace for LOBs and Long Cols
+dialog_table_tools_reorg_resetdict=Reset Dictionary?
+dialog_table_tools_reorg_access_title=Table Access
+dialog_table_tools_reorg_access_no=Allow No Access
+dialog_table_tools_reorg_access_read=Allow Read-Only Access
+dialog_table_tools_reorg_access_readwrite=Allow Read and Write Access
+
+dialog_table_tools_reorgix_title=Reorg Index on Table
+dialog_table_tools_reorgix_access_title=Table Access
+dialog_table_tools_reorgix_access_default=Use Default Access
+dialog_table_tools_reorgix_access_no=Allow No Access
+dialog_table_tools_reorgix_access_read=Allow Read-Only Access
+dialog_table_tools_reorgix_access_readwrite=Allow Read and Write Access
+dialog_table_tools_reorgix_options_title=Options
+dialog_table_tools_reorgix_options_full=Full Index Reorg
+dialog_table_tools_reorgix_options_cleanup_keys=Only Cleanup Pseudo Deleted Keys and Pages
+dialog_table_tools_reorgix_options_cleanup_pages=Only Cleanup Pseudo Deleted Pages
+
+dialog_table_tools_reorgcheck_title=Reorg Check on Table
+
+dialog_table_tools_reorgcheckix_title=Reorg Check on Table Indexes
+
+dialog_table_tools_truncate_title=Truncate Table(s)
+dialog_table_tools_truncate_storage_title=Storage Options
+dialog_table_tools_truncate_storage_reuse=Reuse Storage
+dialog_table_tools_truncate_storage_drop=Drop Storage
+dialog_table_tools_truncate_triggers_title=Delete Triggers Options 
+dialog_table_tools_truncate_triggers_ignore=Ignore Delete Triggers
+dialog_table_tools_truncate_triggers_restrict=Restrict When Delete Triggers
+
+no_ddl_for_system_tables=-- DDL is not available for System Tables
+no_ddl_for_nicknames=-- DDL is not available for Nicknames
+no_ddl_for_nonsql_routines=-- DDL is not available for non-SQL based Routines
+no_ddl_for_spaces_in_name=-- DDL is not available for Object that have a whitespace in their name or Schema
+
+dialog_tools_msg_title=Retrieve SQL Message from SQL Code
+dialog_tools_msg_code=SQL Code? 
+dialog_tools_mes_error_code_title=Error
+dialog_tools_mes_error_code=The SQL Code must be Numeric, eg -551, 803...
+dialog_tools_mes_message=SQL Message
+
