+++ plugins/org.jkiss.dbeaver.ext.db2.ui/src/org/jkiss/dbeaver/ext/db2/ui/views/DB2ConnectionPage.java
@@ -1,0 +1,198 @@
+/*
+ * DBeaver - Universal Database Manager
+ * Copyright (C) 2010-2020 DBeaver Corp and others
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jkiss.dbeaver.ext.db2.ui.views;
+
+import org.eclipse.jface.dialogs.IDialogPage;
+import org.eclipse.jface.resource.ImageDescriptor;
+import org.eclipse.swt.SWT;
+import org.eclipse.swt.events.ModifyEvent;
+import org.eclipse.swt.events.ModifyListener;
+import org.eclipse.swt.layout.GridData;
+import org.eclipse.swt.layout.GridLayout;
+import org.eclipse.swt.widgets.Composite;
+import org.eclipse.swt.widgets.Label;
+import org.eclipse.swt.widgets.Text;
+import org.jkiss.dbeaver.ext.db2.ui.Activator;
+import org.jkiss.dbeaver.ext.db2.ui.internal.DB2Messages;
+import org.jkiss.dbeaver.model.DBPDataSourceContainer;
+import org.jkiss.dbeaver.model.connection.DBPConnectionConfiguration;
+import org.jkiss.dbeaver.ui.ICompositeDialogPage;
+import org.jkiss.dbeaver.ui.UIUtils;
+import org.jkiss.dbeaver.ui.dialogs.connection.ConnectionPageAbstract;
+import org.jkiss.dbeaver.ui.dialogs.connection.DriverPropertiesDialogPage;
+import org.jkiss.utils.CommonUtils;
+
+import java.util.Locale;
+
+/**
+ * DB2ConnectionPage
+ */
+public class DB2ConnectionPage extends ConnectionPageAbstract implements ICompositeDialogPage {
+	private Text hostText;
+	private Text portText;
+	private Text dbText;
+	private Text usernameText;
+
+	private static ImageDescriptor logoImage = Activator.getImageDescriptor("icons/db2_logo.png"); //$NON-NLS-1$
+
+	@Override
+	public void dispose() {
+		super.dispose();
+	}
+
+	@Override
+	public void createControl(Composite composite) {
+		// Composite group = new Composite(composite, SWT.NONE);
+		// group.setLayout(new GridLayout(1, true));
+		setImageDescriptor(logoImage);
+
+		Composite control = new Composite(composite, SWT.NONE);
+		control.setLayout(new GridLayout(1, false));
+		control.setLayoutData(new GridData(GridData.FILL_BOTH));
+		ModifyListener textListener = new ModifyListener() {
+			@Override
+			public void modifyText(ModifyEvent e) {
+				site.updateButtons();
+			}
+		};
+
+		{
+			Composite addrGroup = UIUtils.createControlGroup(control, DB2Messages.db2_connection_page_tab_database, 4,
+					0, 0);
+			GridData gd = new GridData(GridData.FILL_HORIZONTAL);
+			addrGroup.setLayoutData(gd);
+
+			Label hostLabel = UIUtils.createControlLabel(addrGroup, DB2Messages.dialog_connection_host);
+			hostLabel.setLayoutData(new GridData(GridData.HORIZONTAL_ALIGN_END));
+
+			hostText = new Text(addrGroup, SWT.BORDER);
+			gd = new GridData(GridData.FILL_HORIZONTAL);
+			gd.grabExcessHorizontalSpace = true;
+			hostText.setLayoutData(gd);
+			hostText.addModifyListener(textListener);
+
+			Label portLabel = UIUtils.createControlLabel(addrGroup, DB2Messages.dialog_connection_port);
+			gd = new GridData(GridData.HORIZONTAL_ALIGN_END);
+			portLabel.setLayoutData(gd);
+
+			portText = new Text(addrGroup, SWT.BORDER);
+			gd = new GridData(GridData.VERTICAL_ALIGN_BEGINNING);
+			gd.widthHint = 40;
+			portText.setLayoutData(gd);
+			portText.addVerifyListener(UIUtils.getIntegerVerifyListener(Locale.getDefault()));
+			portText.addModifyListener(textListener);
+
+			Label dbLabel = UIUtils.createControlLabel(addrGroup, DB2Messages.dialog_connection_database);
+			dbLabel.setLayoutData(new GridData(GridData.HORIZONTAL_ALIGN_END));
+
+			dbText = new Text(addrGroup, SWT.BORDER);
+			gd = new GridData(GridData.FILL_HORIZONTAL);
+			gd.grabExcessHorizontalSpace = true;
+			gd.horizontalSpan = 3;
+			dbText.setLayoutData(gd);
+			dbText.addModifyListener(textListener);
+		}
+
+		{
+			Composite addrGroup = UIUtils.createControlGroup(control, DB2Messages.db2_connection_page_tab_security, 4,
+					0, 0);
+			GridData gd = new GridData(GridData.FILL_HORIZONTAL);
+			addrGroup.setLayoutData(gd);
+
+			usernameText = UIUtils.createLabelText(addrGroup, DB2Messages.dialog_connection_user_name, "");
+			gd = new GridData(GridData.FILL_HORIZONTAL);
+			gd.widthHint = 200;
+			usernameText.setLayoutData(gd);
+			usernameText.addModifyListener(textListener);
+
+			UIUtils.createEmptyLabel(addrGroup, 2, 1);
+
+			Text passwordText = createPasswordText(addrGroup, DB2Messages.dialog_connection_password);
+			gd = new GridData(GridData.FILL_HORIZONTAL);
+			gd.widthHint = 200;
+			passwordText.setLayoutData(gd);
+			passwordText.addModifyListener(textListener);
+
+			createPasswordControls(addrGroup, 2);
+		}
+
+		createDriverPanel(control);
+		setControl(control);
+	}
+
+	@Override
+	public boolean isComplete() {
+		return hostText != null && portText != null && !CommonUtils.isEmpty(hostText.getText())
+				&& !CommonUtils.isEmpty(portText.getText());
+	}
+
+	@Override
+	public void loadSettings() {
+		super.loadSettings();
+
+		// Load values from new connection info
+		DBPConnectionConfiguration connectionInfo = site.getActiveDataSource().getConnectionConfiguration();
+		if (hostText != null) {
+			hostText.setText(CommonUtils.notEmpty(connectionInfo.getHostName()));
+		}
+		if (portText != null) {
+			if (!CommonUtils.isEmpty(connectionInfo.getHostPort())) {
+				portText.setText(String.valueOf(connectionInfo.getHostPort()));
+			} else if (site.getDriver().getDefaultPort() != null) {
+				portText.setText(site.getDriver().getDefaultPort());
+			} else {
+				portText.setText(""); //$NON-NLS-1$
+			}
+		}
+		if (dbText != null) {
+			dbText.setText(CommonUtils.notEmpty(connectionInfo.getDatabaseName()));
+		}
+		if (usernameText != null) {
+			usernameText.setText(CommonUtils.notEmpty(connectionInfo.getUserName()));
+		}
+		if (passwordText != null) {
+			passwordText.setText(CommonUtils.notEmpty(connectionInfo.getUserPassword()));
+		}
+	}
+
+	@Override
+	public void saveSettings(DBPDataSourceContainer dataSource) {
+		DBPConnectionConfiguration connectionInfo = dataSource.getConnectionConfiguration();
+		if (hostText != null) {
+			connectionInfo.setHostName(hostText.getText().trim());
+		}
+		if (portText != null) {
+			connectionInfo.setHostPort(portText.getText().trim());
+		}
+		if (dbText != null) {
+			connectionInfo.setDatabaseName(dbText.getText().trim());
+		}
+		if (usernameText != null) {
+			connectionInfo.setUserName(usernameText.getText().trim());
+		}
+		if (passwordText != null) {
+			connectionInfo.setUserPassword(passwordText.getText());
+		}
+		super.saveSettings(dataSource);
+	}
+
+	@Override
+	public IDialogPage[] getSubPages(boolean extrasOnly, boolean forceCreate) {
+		return new IDialogPage[] { new DB2ConnectionTracePage(), new DriverPropertiesDialogPage(this), };
+	}
+
+}
