+++ plugins/org.jkiss.dbeaver.ext.db2/META-INF/MANIFEST.MF
@@ -7,9 +7,7 @@ Bundle-Version: 2.1.117.qualifier
 Bundle-Release-Date: 20200504
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
-Bundle-Activator: org.jkiss.dbeaver.ext.db2.Activator
-Require-Bundle: org.eclipse.core.runtime,
- org.eclipse.jface,
+Require-Bundle: org.eclipse.jface,
  org.eclipse.jface.text,
  org.eclipse.ui,
  org.eclipse.ui.views,
@@ -25,3 +23,7 @@ Require-Bundle: org.eclipse.core.runtime,
  org.jkiss.dbeaver.ui.editors.session
 Bundle-ClassPath: .
 Automatic-Module-Name: org.jkiss.dbeaver.ext.db2
+Export-Package: org.jkiss.dbeaver.ext.db2,
+ org.jkiss.dbeaver.ext.db2.editors,
+ org.jkiss.dbeaver.ext.db2.model,
+ org.jkiss.dbeaver.ext.db2.model.dict
